Move GlobalContext to a more appropriate place
